Fix C++20 UTF8 string literals
Since C++20, UTF8 string literals (in the form `u8"blabla"`) have type
`const char8_t[N]`, which does not play nice with llvm::StringRef and
llvm::StringLiteral.

This commit fixes compilation of DJBTest.cpp and in C++20 explicitly
converting those strings literals to regular string literals.

To fix compilation of DJBTest.cpp, we had to define a new UTF8 based
constructor for StringRef and one for StringLiteral.
These cannot be constexpr because const char8_t * cannot be converted at
compile time to const char *.

With these new constructors, the StringRef(data, length) constructor
becomes ambiguous if used with nullptr as first argument, such as
  StringRef(nullptr, 0);
This was done in XCOFFObjectFile.cpp that was fixed to use the default
StringRef constructor.