YAMLParser: allow : and - at plain scalar start
This commit introduces support for having `:` and `-` at the beinning of
a YAML scalar. Right now, YAMLParser disallows them if we are in a flow
sequence (`[ ... ]`), but the YAML specification does not seem to hint
at this limitation.

This enables us to correctly parse YAML that has gone through other YAML
libraries and pyyaml in particular.