Disable DirectoryWatcher tests
These tests started do fail on our CI, for reasons related to the
inotify on CI itself.

The error was the following:

Expected: succeeded
  Actual: failed  (inotify_add_watch() error: No space left on device)

This is for sure not something we've introduced by mistake, since we
haven't changed anything related to DirectoryWatcher in clang.

So it should be safe do disable these tests.