ViewOpGraph: introduce CFG edges printing
Previously, the `ViewOpGraph` pass could only print edges representing
the dataflow, and the control-flow between `Operation`s contained in the
same `Block` (even though this control-flow is very peculiar, since it
is just a straight line connecting all the `Operation`s in a sequential
manner.

In this commit, we implement `Region` control-flow, i.e., control flow
connecting the various `Block`s in a `Region`, connecting the `Block`s
with their successors. It's use is enabled with the
`print-region-control-flow-edges` option.

Since the `ViewOpGraph` pass uses DOT clusters for representing
`Block`s, and edges are not allowed between clusters, we use the first
and last `Operation`s in a `Block`, to represent control-flow between
`Blocks`.