ADT::GenericCycleInfo: Introduce GraphView param
The `GenericCycleInfo` analysis, and all the classes used to perform the
computation, now accept and additional `GraphView` template parameter,
which can be used to specify a view used as the graph on which the
analysis is computed.

This means that we cannot use the `successors` and `predecessors`
helper functions to iterate over the graph, but instead make use of
`llvm::children<GraphView>` and `llvm::children<Inverse<GraphView>>`.